allow choosing custom Rust runtime
