+++ flutter_ffi_plugin/example/native/hub/Cargo.toml
@@ -14,7 +14,12 @@ crate-type = ["lib", "cdylib", "staticlib"]
 [dependencies]
 rinf = "6.15.0"
 prost = "0.13.0"
-tokio = { version = "1", features = ["rt", "sync", "time"] }
-tokio_with_wasm = { version = "0.6.3", features = ["rt", "sync", "time"] }
+tokio = { version = "1", features = ["rt", "sync", "time", "macros"] }
+tokio_with_wasm = { version = "0.7.1", features = [
+    "rt",
+    "sync",
+    "time",
+    "macros",
+] }
 wasm-bindgen = "0.2.93"
 sample_crate = { path = "../sample_crate" }
