+++ flutter_ffi_plugin/example/native/hub/src/lib.rs
@@ -6,6 +6,7 @@ mod messages;
 mod sample_functions;
 
 // use tokio;
+use common::*;
 use tokio_with_wasm::alias as tokio;
 
 rinf::write_interface!();
@@ -15,8 +16,17 @@ rinf::write_interface!();
 // such as `tokio::fs::File::open`.
 // If you really need to use blocking code,
 // use `tokio::task::spawn_blocking`.
-async fn main() {
+#[tokio::main(flavor = "current_thread")]
+async fn main() -> Result<()> {
+    // Spawn the concurrent tasks.
     tokio::spawn(sample_functions::tell_numbers());
     tokio::spawn(sample_functions::stream_fractal());
     tokio::spawn(sample_functions::run_debug_tests());
+
+    // Get the shutdown receiver from Rinf.
+    // This receiver will await a signal from Dart shutdown.
+    let shutdown_receiver = rinf::get_shutdown_receiver()?;
+    shutdown_receiver.await;
+
+    Ok(())
 }
