+++ flutter_ffi_plugin/template/native/hub/src/lib.rs
@@ -15,19 +15,31 @@ rinf::write_interface!();
 // such as `tokio::fs::File::open`.
 // If you really need to use blocking code,
 // use `tokio::task::spawn_blocking`.
+#[tokio::main(flavor = "current_thread")]
 async fn main() {
+    // Spawn the concurrent tasks.
     tokio::spawn(communicate());
+
+    // Get the shutdown receiver from Rinf.
+    // This receiver will await a signal from Dart shutdown.
+    let shutdown_receiver = rinf::get_shutdown_receiver()?;
+    shutdown_receiver.await;
+
+    Ok(())
 }
 
 async fn communicate() -> Result<()> {
     use messages::basic::*;
+
     // Send signals to Dart like below.
     SmallNumber { number: 7 }.send_signal_to_dart();
+
     // Get receivers that listen to Dart signals like below.
     let receiver = SmallText::get_dart_signal_receiver()?;
     while let Some(dart_signal) = receiver.recv().await {
         let message: SmallText = dart_signal.message;
         rinf::debug_print!("{message:?}");
     }
+
     Ok(())
 }
